requirements.txt
apache-airflow>=2.7.1
apache-airflow-providers-apache-spark>=4.1.4
apache-airflow-providers-apache-hdfs>=3.2.1
hdfs>=2.7.0
pandas>=2.0.0
psycopg2-binary>=2.9.9